@fedordikarev 2025-04-02
@bayandin: yes, this PR is follow-up from the todays call with @Bodobolero and @alexanderlaw with the idea to check first if we will get reproducible results with new runners type, before we go any further with that hypothesis.