@Bodobolero 2025-04-14
> Are these significantly slower than production storage nodes? Should we use hardware that has roughly the same single-threaded/IO performance as production, so that the numbers are transferrable and we have a rough idea of the equivalent performance in production?

We are now testing setting the governor to "performance" profile which achieves approximately same speed as the previously used runners while still achieving consistent performance:

Table of runtimes for the same commit on different runners in [run](https://github.com/neondatabase/neon/actions/runs/14417589789) 

| Run | Benchmarks (1) | Benchmarks (2) |Benchmarks (3) |Benchmarks (4) | Benchmarks (5) |
|--------|--------|---------|---------|---------|---------|
| 1 | 1950.37s | 6374.55s |  3646.15s |  4149.48s |  2330.22s | 
| 2 | - | 6369.27s |  3666.65s |  4162.42s |  2329.23s | 
| Delta % |  - |  0,07 %  | 0,5 %   |   0,3 % | 0,04 %   |
| with governor performance | 1519.57s |  4131.62s |  - | -  |  - |
| second run gov. perf. | 1513.62s |  - |  - | -  |  - |
| speedup gov. performance | 22 % |  35 % |  - | -  |  - |
| current desktop class hetzner runners (main) | 1487.10s |  3699.67s |  - | -  |  - |