@Bodobolero 2025-04-14
We initially deployed 2 runners and for now decided to [deploy 3 more](https://github.com/neondatabase/cloud/issues/27551) before we merge this and also automate the governor to use "performance" instead of schedutil (when deployed and after reboot).